refactor: toolbars
